Certificates modified

+++ app/Http/Controllers/aResumesController.php
@@ -22,21 +22,32 @@ class aResumesController extends Controller {
 	/**
 	 * Display the resume of the Profile.
 	 *
-	 * @param  int $id 	profile_id
 	 * @return Response
 	 */
-	public function index($id)
+	public function index()
 	{
 
-		$profile = Profile::findOrFail($id);
+        if (! $user = JWTAuth::parseToken()->toUser()) {
+            return response()->json(['error' => 'user_not_found'], 400);
+        }
+
+        try {
+          $profile = Profile::byuser_id($user->id)->firstOrFail();
+          $resume = $profile->resume;
+        } catch(\Exception $e) {
+            return response()->json(['error' => $e->getMessage()], $e->getCode());
+        } catch (QueryException $e){
+            $errorCode = $e->errorInfo[1];
+            return response()->json(['error' => $error_code], 500);
+        }        
+
 		if(!$profile->resume)
 			return response()->json(['resume_not_found'], 404);
 		else
 		{
-			$resume = collect([]);
-			$resume->push(['id' => $profile->resume, 'cv' => $profile->resume->cv]);
+			$cv = $resume->cv;
+			return response()->json(compact('cv'));	
 		}
-		return response()->json(compact('resume'));		
 	}
 
 	/**
@@ -50,36 +61,46 @@ class aResumesController extends Controller {
 	}
 
 	/**
-	 * Store a newly created resource in storage.
+	 * Create a Resume of Profile and the CV record of Resume
+	 * with text from input
 	 *
-	 * @param  int $id 	profile_id
 	 * @param Request $request
 	 * @return Response
 	 */
-	public function store($id, Request $request)
+	public function store(Request $request)
 	{
-		$profile = Profile::findOrFail($id);
+        if (! $user = JWTAuth::parseToken()->toUser()) {
+            return response()->json(['error' => 'user_not_found'], 400);
+        }
 
-		if(!$profile->resume){
+        try {
+          $profile = Profile::byuser_id($user->id)->firstOrFail();
+          $resume = $profile->resume;
+        } catch(\Exception $e) {
+            return response()->json(['error' => $e->getMessage()], $e->getCode());
+        } catch (QueryException $e){
+            $errorCode = $e->errorInfo[1];
+            return response()->json(['error' => $error_code], 500);
+        }        
+
+
+		if(!$profile->resume) {
 			$resume = new Resume;
 			$resume->cv = $request->input('cv');
-			
-			try {
-				//$resume->save();
-				$profile->resume()->save($resume);			
-			} catch(\Exception $e) {
-			    return response()->json(["error" => $e->getCode()], 500);
-			}
-
-			return response()->json(['success profile'. $profile->id], 200);
 		}
 		else
 		{
-			return response()->json(['resume exists, go to update'], 404);
+			$resume->cv = $request->input('cv');
 		}
 
-
-
+		try {
+		  	$profile->resume()->save($resume);			
+		} catch(\Exception $e) {
+            return response()->json(['error' => $e->getMessage()], $e->getCode());
+		}
+		
+		return response()->json(['success profile'. $profile->id], 200);
+			
 	}
 
 	/**
@@ -90,7 +111,28 @@ class aResumesController extends Controller {
 	 */
 	public function show($id)
 	{
-		//
+        if (! $user = JWTAuth::parseToken()->toUser()) {
+            return response()->json(['error' => 'user_not_found'], 400);
+        }
+
+        try {
+          $profile = Profile::byuser_id($user->id)->firstOrFail();
+          $resume = $profile->resume;
+        } catch(\Exception $e) {
+            return response()->json(['error' => $e->getMessage()], $e->getCode());
+        } catch (QueryException $e){
+            $errorCode = $e->errorInfo[1];
+            return response()->json(['error' => $error_code], 500);
+        }        
+
+		if(!$profile->resume)
+			return response()->json(['resume_not_found'], 404);
+		else
+		{
+			$cv = $resume->cv;
+			return response()->json(compact('cv'));	
+		}
+
 	}
 
 	/**
@@ -105,16 +147,26 @@ class aResumesController extends Controller {
 	}
 
 	/**
-	 * Update the specified resource in storage.
+	 * Update the CV record of Resume of current Profile.
 	 *
-	 * @param  int $id 	profile_id
 	 * @param Request $request
 	 * @return Response
 	 */
-	public function update($id, Request $request)
+	public function update(Request $request)
 	{
-		$profile = Profile::findOrFail($id);
-		$resume = $profile->resume;
+        if (! $user = JWTAuth::parseToken()->toUser()) {
+            return response()->json(['error' => 'user_not_found'], 400);
+        }
+
+        try {
+          $profile = Profile::byuser_id($user->id)->firstOrFail();
+          $resume = $profile->resume;
+        } catch(\Exception $e) {
+            return response()->json(['error' => $e->getMessage()], $e->getCode());
+        } catch (QueryException $e){
+            $errorCode = $e->errorInfo[1];
+            return response()->json(['error' => $error_code], 500);
+        }
 
 		if(!$profile->resume)
 			return response()->json(['resume_not_created'], 404);
@@ -125,33 +177,43 @@ class aResumesController extends Controller {
 			try {
 				$profile->resume()->save($resume);			
 			} catch(\Exception $e) {
-		    	return response()->json(["error" => $e->getCode()], 500);
+            return response()->json(['error' => $e->getMessage()], $e->getCode());
 			}
 		
 			return response()->json(['success profile'. $profile->id], 200);
 		}
-
 	}
 
 	/**
 	 * Remove the specified resource from storage.
 	 *
-	 * @param  int $id 	profile_id
 	 * @return Response
 	 */
-	public function destroy($id)
+	public function destroy()
 	{
-		$profile = Profile::findOrFail($id);
+        if (! $user = JWTAuth::parseToken()->toUser()) {
+            return response()->json(['error' => 'user_not_found'], 400);
+        }
+
+        try {
+          $profile = Profile::byuser_id($user->id)->firstOrFail();
+          $resume = $profile->resume;
+        } catch(\Exception $e) {
+            return response()->json(['error' => $e->getMessage()], $e->getCode());
+        } catch (QueryException $e){
+            $errorCode = $e->errorInfo[1];
+            return response()->json(['error' => $error_code], 500);
+        }
+
 		File::delete($profile->resume->file_path);
 		$profile->resume()->delete();
 
-		$profile = Profile::findOrFail($id);
 		return response()->json(['success profile'. $profile->id], 200);
 		
 	}
 
 	/**
-	 * Update the specified resource in storage.
+	 * Update the file_path record of Resume of current Profile and upload new file
 	 *
 	 * @param Request $request
 	 * @return Response
@@ -179,16 +241,15 @@ class aResumesController extends Controller {
 		$file_name = 'cv'. $profile->id . '.' . $request->file('file')->getClientOriginalExtension();
 		$resume->file_path = public_path() . "\\resumes\\" . $file_name;
 		$request->file('file')->move('resumes', $file_name);
+		  //daca fisierul exista se suprascrie
 
 		try {
-		  //daca fisierul exista se suprascrie
 		  $profile->resume()->save($resume);			
 		} catch(\Exception $e) {
             return response()->json(['error' => $e->getMessage()], $e->getCode());
 		}
 		
 		return response()->json(['success profile'. $profile->id], 200);
-
 	}
 
 }
